anime-service/src/main/java/academy/devdojo/animeservice/repository/AnimeHardCodedRepository.java
package academy.devdojo.animeservice.repository;

import academy.devdojo.animeservice.domain.Anime;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;

import java.util.List;
import java.util.Optional;

@Repository
@RequiredArgsConstructor
public class AnimeHardCodedRepository {

    private final AnimeData animeData;

    public List<Anime> listAll() {
        return animeData.getAnimes();
    }

    public List<Anime> findByName(String name) {
        return name == null ? animeData.getAnimes() :
                animeData.getAnimes().stream()
                        .filter(anime -> anime.getName().equalsIgnoreCase(name))
                        .toList();
    }

    public Optional<Anime> findById(Long id) {
        return animeData.getAnimes().stream()
                .filter(anime -> anime.getId().equals(id))
                .findFirst();
    }

    public Anime save(Anime anime) {
        animeData.getAnimes().add(anime);
        return anime;
    }

    public void delete(Anime anime) {
        animeData.getAnimes().remove(anime);
    }

    public void update(Anime anime) {
        animeData.getAnimes().remove(anime);
        animeData.getAnimes().add(anime);
    }
}
